hot fix 10/12/2023

+++ danny-store/src/main/java/danny/store/dannystore/domain/dto/CartDtoAndAmount.java
@@ -6,5 +6,5 @@ import java.util.List;
 @Data
 public class CartDtoAndAmount {
     private List<CartDto> cartDtoList;
-    private Float totalAmount;
+    private Long totalAmount;
 }
